requirements.txt
watchdog
streamlit
pandas
numpy
plotly
requests
tweepy
python-dotenv
praw